api-tests/requirements.txt
pytest>=7.4.0
httpx>=0.25.0
pydantic>=2.5.0
allure-pytest>=2.13.0
pytest-bdd>=7.0.0
playwright>=1.40.0
pytest-playwright>=0.4.0
locust>=2.20.0
